✨ Add websocket server (#203)
* ✨ Add websocket server

The server is versioned as well so respective server versions can send different
data. When a websocket connects to the v1 route it will recieve a message
anytime the v1/collect endpoint updates the database.

* 🐛 Properly handle websocket server versioning

When new websocket server versions are created, they all cannot share the same http
server. To have the websocket servers share a http server, we have to handle the
upgrade even manually.