♻️🏗 Remove Webpack & Karma, replace with Parcel & Cypress (#757)
* ♻🏗️ Refactor to use Parcel 2 over Webpack

'cause I wanted to use it and also too lazy to upgrade to the next major version
of webpack

* ♻️🏗 Remove Karma tests, use Cypress

This is because Parcel isn't really going to work well with Karma. As much as I
love Karma, I don't want to sink my time into making it work well with
Parcel. The wins from using Cypress is these are E2E tests hitting the local
server (I've been trolled by dep updates in the past because of the mocked API)
and it makes us make interactor.js work with Cypress.

* 🏗 Add GitHub actions